refactored for static pip and pyparsing install
